Apply same pattern as MMR
